build: resolve tsconfig error
The tsconfig was set up to emit outputs, but not exclude the directory
containing those outputs from future compilation, so it was giving off
compiler errors in the editor. This could also be solved by changing the
`outDir` to "./build", but it was not immediately apparent what use a
standalone app had for emitted TypeScript types.